rename: shell.py -> shell_tester.py
