+++ shell/shell_tester.py
